compute: tokenize linear join closure
This commit adds a shutdown token check to the closure we pass to the
differential join operator. When the dataflow is shutting down, this
makes the join closure drain all input data, rather than processing it.
As a result, differential join operators shut down faster an emit less
data, which in turn speeds up shutdown of downstream operators.

Unfortunately, the new shutdown logic interferes with the fueling of the
differential join operator. Fuel is consumed based on the number of
updates emitted. When the token is dropped, the join closure stops
producing updates, which means the operator stops consuming fuel, so it
does not yield anymore until it has drained all its inputs. If there are
many inputs left, the replica may not accept commands for potentially
quite a long time.