compute: synchronizing shutdown token
This commit replaces the compute shutdown token, which previously was an
`Rc<()>`, with a `Button`. The `Button` synronizes between workers and
only reports as having been "pressed" when all workers have dropped
their instance of it. This avoids race conditions caused by different
workers dropping their shutdown tokens at different times, making others
observe partial/inconsistent results in the process.

This commit also renames the existing `ShutdownToken` type to
`ShutdownProbe`, to make it clearer that this type encapsulates the
observing side, not the side that will drop the token.