k8s: reduce client timeouts
The default timeouts for the k8s client are:

 * connect:  30s
 * read:    295s
 * write:   295s

We always connect to a k8s cluster that's running in the same region as
we do, so there is no reason we'd need such long timeouts.

Reducing them lets us react faster to network issues, by retrying the
request. In particular, it seems to fix a test flake where sometimes the
command to create the service for a new cluster would get stuck until
the test timed out.