cluster: remove the activator back-channel
Now that the compute command channel dataflow is persistent, rather than
recreated on every client connection, there is no need to send back
connection-specific activators from the workers to the `ClusterClient`.
Instead the `ClusterClient` can simply unpark the threads running the
client workers, which it can get at through the `TimelyContainer`.

Removing the activator channel removes a potential source of cluster
deadlocks as now the `ClusterClient`'s progress isn't dependent on the
liveness of the workers anymore. It also simplifies the code a bunch.