k8s-orchestrator: create a statefulset per process
This commit changes the k8s orchestrator to create a separate
statefulset for each process requested, instead of using a single
statefulset with multiple replicas.

This is to allow us to supply different configurations to different
cluster processes. In particular, we want to be able to provide each
process with its index through a command line argument.